renamed general to dispatcher
