+++ src/command_run/anilist/level.rs
@@ -9,7 +9,7 @@ use crate::command_run::anilist::user::get_user_data;
 use crate::constant::{COMMAND_SENDING_ERROR, OPTION_ERROR};
 use crate::error_enum::AppError;
 use crate::lang_struct::anilist::level::load_localization_level;
-use crate::database::general::data::get_registered_user;
+use crate::database::dispatcher::data_dispatch::get_registered_user;
 
 pub async fn run(
     options: &[CommandDataOption],
