+++ src/command_run/general/module.rs
@@ -6,7 +6,7 @@ use serenity::all::{
 use crate::constant::{COLOR, COMMAND_SENDING_ERROR};
 use crate::error_enum::AppError;
 use crate::lang_struct::general::module::load_localization_module_activation;
-use crate::database::general::data::{
+use crate::database::dispatcher::data_dispatch::{
     get_data_module_activation_status, set_data_module_activation_status,
 };
