--- src/database/general/mod.rs
@@ -1,3 +1,0 @@
-pub mod cache;
-pub mod data;
-pub mod sql;
