+++ src/main.rs
@@ -15,7 +15,7 @@ use crate::components::components_dispatch::components_dispatching;
 use crate::constant::ACTIVITY_NAME;
 use crate::game_struct::steam_game_id_struct::get_game;
 use crate::logger::{create_log_directory, init_logger, remove_old_logs};
-use crate::database::general::sql::init_sql_database;
+use crate::database::dispatcher::init_dispatch::init_sql_database;
 
 mod activity;
 mod anilist_struct;
